op access to quarries; guide book; chests above bars
Server OP's can now interact with and use all quarries.

Added a guide book - Type "/quarrycraft guide" to get one. (Players are told this upon joining the server)

Chests can now also be placed on top of the iron bars (They must be directly above the iron bars, if you'd like to use hoppers to extract items you'll probably want to make it a double chest to reach the bottom).